Adding comments and some minor restructuring | 1
